trying to fix up tests

+++ lib/puppet/provider/windowsfeature/default.rb
@@ -4,8 +4,9 @@ Puppet::Type.type(:windowsfeature).provide(:default) do
   # We don't support 1.8.7 officially, but lets be nice and not cause errors
   # rubocop:disable Style/HashSyntax
 
+  # windows only
   confine :kernel => :windows
-
+  # powershell, powershell, powershell.  where to find it.
   commands :ps =>
     if File.exist?("#{ENV['SYSTEMROOT']}\\sysnative\\WindowsPowershell\\v1.0\\powershell.exe")
       "#{ENV['SYSTEMROOT']}\\sysnative\\WindowsPowershell\\v1.0\\powershell.exe"
@@ -14,33 +15,36 @@ Puppet::Type.type(:windowsfeature).provide(:default) do
     else
       'powershell.exe'
     end
-  
-  def initialize(value={})
-    super(value)
-    @property_flush = {}
-    @install   = 'Install-WindowsFeature'
-  end
 
   def self.instances
+    # an array to store feature hashes
     features = []
-    result = if Facter.value(:kernelmajversion) == '6.1'
-      ps('Import-Module ServerManager; Get-WindowsFeature | ConvertTo-XML -As String -Depth 4 -NoTypeInformation')
-    else
-      ps('Get-WindowsFeature | ConvertTo-XML -As String -Depth 4 -NoTypeInformation')
-    end
+    # set windows 2008 to true/False
+    win2008 = Facter.value(:kernelmajversion) == '6.1'
+    # if win2008 import ServerManager module
+    result = if win2008 == true
+               ps('Import-Module ServerManager; Get-WindowsFeature | ConvertTo-XML -As String -Depth 4 -NoTypeInformation')
+             else
+               ps('Get-WindowsFeature | ConvertTo-XML -As String -Depth 4 -NoTypeInformation')
+             end
+    # create the XML document and parse the objects
     xml = Document.new result
     xml.root.each_element do |object|
-      name   = object.elements["Property[@Name='Name']"].text.downcase
-      state  = if object.elements["Property[@Name='Installed']"].text == 'False'
-                 :absent
-               elsif object.elements["Property[@Name='Installed']"].text == 'True'
-                 :present
-               end
+      # get the name and install state of the windows feature
+      name  = object.elements["Property[@Name='Name']"].text.downcase
+      state = if object.elements["Property[@Name='Installed']"].text == 'False'
+                :absent
+              elsif object.elements["Property[@Name='Installed']"].text == 'True'
+                :present
+              end
+      # put name and state into a hash
       feature_hash = {
-        :name => name, :ensure => state,
+        :ensure => state, :name => name
       }
+      # push hash to feature array
       features.push(feature_hash)
     end
+    # map the feature array
     features.map do |feature|
       new(feature)
     end
@@ -60,38 +64,42 @@ Puppet::Type.type(:windowsfeature).provide(:default) do
   end
 
   def create
+    # an array called array
     array = []
-    if Facter.value(:kernelmajversion) == '6.1'
-      array << "Import-Module ServerManager; Add-WindowsFeature #{resource[:name]}"
-    else
-      array << "Install-WindowsFeature #{resource[:name]}"
+    # if it is windows 2008 let's just call it that
+    win2008 = Facter.value(:kernelmajversion) == '6.1'
+    # set the install line
+    array << "Import-Module ServerManager; Add-WindowsFeature #{resource[:name]}" if win2008 == true
+    array << "Install-WindowsFeature #{resource[:name]}" if win2008 == false
+    # add restart, subfeatures and a source optionally
+    array << '-IncludeAllSubFeature' if @resource[:installsubfeatures] == true
+    array << '-Restart' if @resource[:restart] == true
+    array << "-Source #{resource['source']}" if @resource[:source]== false
+    # raise an error if 2008 tried to install mgmt tools
+    if @resource[:installmanagementtools] == true && win2008 == true
+      raise Puppet::Error, 'installmanagementtools can only be used with Windows 2012 and above'
     end
-    array << '-IncludeAllSubFeature' if @resource[:installsubfeatures]
-    array << '-Restart' if @resource[:restart]
-    array << "-Source #{resource['source']}" if @resource[:source]
-    errormsg = 'installmanagementtools can only be used with Windows 2012 and above'
-    if Facter.value(:kernelmajversion) == '6.1'
-      raise Puppet::Error, errormsg if @resource[:installmanagementtools]
-    else
-      array << '-IncludeManagementTools' if @resource[:installmanagementtools]
-    end
-    Puppet.notice array
-    Puppet.debug "Powershell create command is '#{array}''"
+    # install management tools
+    array << '-IncludeManagementTools' if @resource[:installmanagementtools] == true && win2008 == false
+    # show the created ps string, get the result, show the result (debug)
+    Puppet.debug "Powershell create command is '#{array}'"
     result = ps(array.join(' '))
-    Puppet.debug "Powershell create response was #{result}"
+    Puppet.debug "Powershell create response was '#{result}'"
   end
 
   def destroy
+    # an array called array
     array = []
-    if Facter.value(:kernelmajversion) == '6.1'
-      array << "Import-Module ServerManager; Remove-WindowsFeature #{resource[:name]}"
-    else
-      array << "Uninstall-WindowsFeature #{resource[:name]}"
-    end
-    array << '-Restart' if @resource[:restart]
-    Puppet.debug "Powershell destroy command is '#{array}''"
+    # if it is windows 2008 let's just call it that
+    win2008 = Facter.value(:kernelmajversion) == '6.1'
+    # set the uninstall line
+    array << "Import-Module ServerManager; Remove-WindowsFeature #{resource[:name]}" if win2008 == true
+    array << "Uninstall-WindowsFeature #{resource[:name]}" if win2008 == false
+    # add the restart flag optionally
+    array << '-Restart' if @resource[:restart] == true
+    # show the created ps string, get the result, show the result (debug)
+    Puppet.debug "Powershell destroy command is '#{array}'"
     result = ps(array.join(' '))
-    Puppet.debug "Powershell destroy response was #{result}"
+    Puppet.debug "Powershell destroy response was '#{result}'"
   end
-
 end
